// src/beneficios/DTOs/alloyal-organization-highlight-online.dto.ts
export class AlloyalOrganizationHighlightOnlineDto {
  id: number;
  name: string;
  cover_picture: string;
  background_picture: string;
  best_discount_percent: number;
  category_name: string;
  instagram_url: string;
  facebook_url: string;
  twitter_url: string;
  cashback_percent: number;
  cashback_text: string;
  discount_text: string;
  description: string;
}